unify contract and script call handlers
