+++ packages/fuels-code-gen/src/program_bindings/abigen/bindings/script.rs
@@ -108,9 +108,10 @@ fn expand_fn(fn_abi: &FullABIFunction) -> Result<TokenStream> {
     let mut generator = FunctionGenerator::new(fn_abi)?;
 
     let arg_tokens = generator.tokenized_args();
+    let original_output_type = generator.output_type();
     let body = quote! {
             let encoded_args = ::fuels::core::codec::ABIEncoder::new(self.encoder_config).encode(&#arg_tokens);
-            ::fuels::programs::script_calls::ScriptCallHandler::new(
+            ::fuels::programs::call_handler::CallHandler::new_script_call(
                 self.binary.clone(),
                 encoded_args,
                 self.account.clone(),
@@ -118,12 +119,8 @@ fn expand_fn(fn_abi: &FullABIFunction) -> Result<TokenStream> {
             )
     };
 
-    let original_output_type = generator.output_type();
-
     generator
-        .set_output_type(
-            quote! {::fuels::programs::script_calls::ScriptCallHandler<T, #original_output_type> },
-        )
+        .set_output_type(quote! {::fuels::programs::call_handler::CallHandler<T, #original_output_type, ::fuels::programs::calls::ScriptCall> })
         .set_docs(fn_abi.doc_strings()?)
         .set_body(body);
 
@@ -189,10 +186,10 @@ mod tests {
         let expected = quote! {
             #[doc = "This is a doc string"]
             #[doc = "This is another doc string"]
-            pub fn main(&self, bimbam: ::core::primitive::bool) -> ::fuels::programs::script_calls::ScriptCallHandler<T, ()> {
+            pub fn main(&self, bimbam: ::core::primitive::bool) -> ::fuels::programs::call_handler::CallHandler<T, (), ::fuels::programs::call::ScriptCall> {
                 let encoded_args=::fuels::core::codec::ABIEncoder::new(self.encoder_config)
                     .encode(&[::fuels::core::traits::Tokenizable::into_token(bimbam)]);
-                 ::fuels::programs::script_calls::ScriptCallHandler::new(
+                 ::fuels::programs::call_handler::CallHandler::new_script_call(
                     self.binary.clone(),
                     encoded_args,
                     self.account.clone(),
