+++ packages/fuels-programs/src/calls.rs
@@ -1,0 +1,9 @@
+mod callable;
+mod contract;
+pub mod receipt_parser;
+mod script;
+pub mod utils;
+
+pub use callable::*;
+pub use contract::*;
+pub use script::*;
