+++ packages/fuels-programs/src/calls/receipt_parser.rs
