+++ packages/fuels-programs/src/lib.rs
@@ -1,6 +1,4 @@
-pub mod call_response;
-pub mod call_utils;
+pub mod call_handler;
+pub mod calls;
 pub mod contract;
-pub mod receipt_parser;
-pub mod script_calls;
-mod submit_response;
+pub mod responses;
