+++ packages/fuels-programs/src/responses.rs
@@ -1,0 +1,5 @@
+mod call;
+mod submit;
+
+pub use call::*;
+pub use submit::*;
