+++ packages/fuels-programs/src/responses/call.rs
@@ -7,14 +7,12 @@ use fuels_core::{
     types::errors::Result,
 };
 
-/// [`FuelCallResponse`] is a struct that is returned by a call to the contract or script. Its value
+/// [`CallResponse`] is a struct that is returned by a call to the contract or script. Its value
 /// field holds the decoded typed value returned by the contract's method. The other field holds all
 /// the receipts returned by the call.
-/// The name is `FuelCallResponse` instead of `CallResponse` because it would be ambiguous with the
-/// `CALL` opcode.
 #[derive(Debug)]
 // ANCHOR: fuel_call_response
-pub struct FuelCallResponse<D> {
+pub struct CallResponse<D> {
     pub value: D,
     pub receipts: Vec<Receipt>,
     pub gas_used: u64,
@@ -23,7 +21,7 @@ pub struct FuelCallResponse<D> {
 }
 // ANCHOR_END: fuel_call_response
 
-impl<D> FuelCallResponse<D> {
+impl<D> CallResponse<D> {
     /// Get the gas used from ScriptResult receipt
     fn get_gas_used(receipts: &[Receipt]) -> u64 {
         receipts
